+++ main.tf
@@ -10,6 +10,16 @@ locals {
   template_cluster          = file("${path.module}/conf/nomad/presto.hcl")
   consul_connect_plugin_uri = "${var.consul_connect_plugin_artifact_source}/${var.consul_connect_plugin_version}/presto-consul-connect-${var.consul_connect_plugin_version}-jar-with-dependencies.jar"
   node_types                = var.coordinator ? ["coordinator", "worker"] : ["worker"]
+
+  vault_provider = var.vault_secret.use_vault_provider || var.minio_vault_secret.use_vault_provider
+  vault_kv_policy_name_set = jsonencode(
+    distinct(
+      concat(
+        [var.vault_secret.vault_kv_policy_name],
+        [var.minio_vault_secret.vault_kv_policy_name]
+      )
+    )
+  )
 }
 
 data "template_file" "template_nomad_job_presto" {
@@ -21,12 +31,16 @@ data "template_file" "template_nomad_job_presto" {
     datacenters    = local.datacenters
     namespace      = var.nomad_namespace
 
+    # policies if vault secret kv provider is ON
+    use_vault_provider = local.vault_provider
+    vault_policy_array = local.vault_kv_policy_name_set
+
     # presto
     service_name              = var.service_name
     node_types                = jsonencode(local.node_types)
     local_docker_image        = var.local_docker_image
     shared_secret_user        = var.shared_secret_user
-    use_vault_secret_provider = var.vault_secret.use_vault_secret_provider
+    use_vault_secret_provider = var.vault_secret.use_vault_provider
     vault_kv_policy_name      = var.vault_secret.vault_kv_policy_name
     vault_kv_path             = var.vault_secret.vault_kv_path
     vault_kv_secret_key_name  = var.vault_secret.vault_kv_secret_key_name
@@ -53,14 +67,21 @@ data "template_file" "template_nomad_job_presto" {
     consul_connect_plugin_uri = local.consul_connect_plugin_uri
 
     #hivemetastore
-    hivemetastore_service_name = var.hivemetastore.service_name
-    hivemetastore_port         = var.hivemetastore.port
+    hivemetastore_service_name = var.hivemetastore_service.service_name
+    hivemetastore_port         = var.hivemetastore_service.port
 
     # minio
-    minio_service_name = var.minio.service_name
-    minio_port         = var.minio.port
-    minio_access_key   = var.minio.access_key
-    minio_secret_key   = var.minio.secret_key
+    minio_service_name = var.minio_service.service_name
+    minio_port         = var.minio_service.port
+    minio_access_key   = var.minio_service.access_key
+    minio_secret_key   = var.minio_service.secret_key
+
+    ## if creds are provided by vault
+    minio_use_vault_provider       = var.minio_vault_secret.use_vault_provider
+    minio_vault_kv_policy_name     = var.minio_vault_secret.vault_kv_policy_name # This var is appended to local.vault_kv_policy_name_set
+    minio_vault_kv_path            = var.minio_vault_secret.vault_kv_path
+    minio_vault_kv_access_key_name = var.minio_vault_secret.vault_kv_access_key_name
+    minio_vault_kv_secret_key_name = var.minio_vault_secret.vault_kv_secret_key_name
   }
 }